Add rule combiner and associated tests
It seems silly to produce policies with a ton of rules all referencing
different keys of the same label, as these will become separate iptables
rules and ipsets. Instead, we should combine these rules to use
MatchExpressions and the 'in' operator. I have not attempted to cover
complex cases eg with namespace selectors or ip ranges or multiple labels.